Restructure Exception system
