+++ python-package/basedosdados/download/download.py
@@ -10,7 +10,11 @@ import re
 import pandas as pd
 from basedosdados.upload.base import Base
 from functools import partialmethod
-from basedosdados.validation.exceptions import BaseDosDadosException
+from basedosdados.exceptions import (
+    BaseDosDadosException, BaseDosDadosAccessDeniedException,
+    BaseDosDadosAuthorizationException, BaseDosDadosInvalidProjectIDException,
+    BaseDosDadosNoBillingProjectIDException
+)
 from pandas_gbq.gbq import GenericGBQException
 
 
@@ -165,46 +169,15 @@ def read_sql(query, billing_project_id=None, from_file=False, reauth=False):
 
     except GenericGBQException as e:
         if "Reason: 403" in str(e):
-            msg = (
-                "\nAre you sure you are using the right `billing_project_id`?"
-                "\nYou must use the Project ID available in your Google Cloud"
-                " console home page at https://console.cloud.google.com/home/dashboard"
-                "\nIf you still don't have a Google Cloud Project, you have to "
-                "create one.\n"
-                "You can set one up by following these steps: \n"
-                "1. Go to this link https://console.cloud.google.com/projectselector2/home/dashboard\n"
-                "2. Agree with Terms of Service if asked\n"
-                "3. Click in Create Project\n"
-                "4. Put a cool name in your project\n"
-                "5. Hit create\n"
-                "6. Rerun this command with the flag `reauth=True`. \n"
-                "   Like `read_table('br_ibge_pib', 'municipios', "
-                "billing_project_id=<YOUR_PROJECT_ID>, reauth=True)`"
-            )
-            raise BaseDosDadosException(msg) from e
+            raise BaseDosDadosAccessDeniedException
         
         elif re.match("Reason: 400 POST .* [Pp]roject[ ]*I[Dd]", str(e)):
-            msg = (
-                "\nYou are using an invalid `billing_project_id`.\nMake sure "
-                "you set it to the Project ID available in your Google Cloud"
-                " console home page at https://console.cloud.google.com/home/dashboard"
-            )
-            raise BaseDosDadosException(msg) from e
+            raise BaseDosDadosInvalidProjectIDException
 
         raise
 
     except PyDataCredentialsError as e:
-        msg = (
-            "\nAre you sure you did the authorization process correctly?\n"
-            "If you were given the option to enter an authorization code, "
-            "please try again and make sure you are entering the right one."
-            "\nYou can try again by rerunning this command with the flag "
-            "`reauth=True`. \n\tLike `read_table('br_ibge_pib', 'municipios',"
-            " billing_project_id=<YOUR_PROJECT_ID>, reauth=True)`"
-            "\nThen you can click at the provided link and get the right "
-            "authorization code."
-        )
-        raise BaseDosDadosException(msg) from e
+        raise BaseDosDadosAuthorizationException
 
     except (OSError, ValueError) as e:
         exc_from_no_billing_id = (
@@ -212,24 +185,7 @@ def read_sql(query, billing_project_id=None, from_file=False, reauth=False):
             "reading from stdin while output is captured" in str(e)
         )
         if exc_from_no_billing_id:
-            msg = (
-                "\nWe are not sure which Google Cloud project should be billed.\n"
-                "First, you should make sure that you have a Google Cloud project.\n"
-                "If you don't have one, set one up following these steps: \n"
-                "\t1. Go to this link https://console.cloud.google.com/projectselector2/home/dashboard\n"
-                "\t2. Agree with Terms of Service if asked\n"
-                "\t3. Click in Create Project\n"
-                "\t4. Put a cool name in your project\n"
-                "\t5. Hit create\n"
-                "\n"
-                "Copy the Project ID, (notice that it is not the Project Name)\n"
-                "Now, you have two options:\n"
-                "1. Add an argument to your function poiting to the billing project id.\n"
-                "   Like `bd.read_table('br_ibge_pib', 'municipios', billing_project_id=<YOUR_PROJECT_ID>)`\n"
-                "2. You can set a project_id in the environment by running the following command in your terminal: `gcloud config set project <YOUR_PROJECT_ID>`.\n"
-                "   Bear in mind that you need `gcloud` installed."
-            )
-            raise BaseDosDadosException(msg) from e
+            raise BaseDosDadosNoBillingProjectIDException
         raise
