--- python-package/basedosdados/validation/exceptions.py
@@ -1,2 +1,0 @@
-class BaseDosDadosException(Exception):
-    """Exclusive Exception from Base dos Dados"""
